Reorganizeb1 shimming files
